requirements.txt
Django==1.5.4
PIL==1.1.7
dj-database-url==0.2.2
dj-static==0.0.5
django-bootstrap-toolkit==2.15.0
django-bootstrap3==1.0.0
django-crispy-forms==1.4.0
django-localflavor==1.0
django-parsley==0.1
django-registration==1.0
django-taggit==0.11.2
django-toolbelt==0.0.1
gunicorn==18.0
ipython==1.1.0
oauthlib==0.6.0
psycopg2==2.5.1
python-openid==2.2.5
python-social-auth==0.1.13
requests==1.2.3
requests-oauthlib==0.3.3
six==1.4.1
static==0.4
wsgiref==0.1.2
